move settings manager to core widgets and add author
