+++ release/scripts/mgear/core/widgets.py
@@ -231,7 +231,6 @@ class DragQListView(QtWidgets.QListView):
         self.ignore_self = True
 
     def mouseMoveEvent(self, event):
-
         mimeData = QtCore.QMimeData()
         mimeData.setText("%d,%d" % (event.x(), event.y()))
 
@@ -272,7 +271,6 @@ class DragQListView(QtWidgets.QListView):
 
 
 class CollapsibleHeader(QtWidgets.QWidget):
-
     COLLAPSED_PIXMAP = QtGui.QPixmap(":teRightArrow.png")
     EXPANDED_PIXMAP = QtGui.QPixmap(":teDownArrow.png")
 
@@ -366,14 +364,17 @@ class CollapsibleWidget(QtWidgets.QWidget):
 # Helper functions
 ##################
 
-def create_button(size=17,
-                  text=None,
-                  icon=None,
-                  toggle_icon=None,
-                  icon_size=None,
-                  toolTip=None,
-                  width=None,
-                  setMax=True):
+
+def create_button(
+    size=17,
+    text=None,
+    icon=None,
+    toggle_icon=None,
+    icon_size=None,
+    toolTip=None,
+    width=None,
+    setMax=True,
+):
     """Create and configure a QPushButton
 
     Args:
@@ -410,13 +411,11 @@ def create_button(size=17,
         # button.setIconSize(QtCore.QSize(icon_size, icon_size))
 
     if toggle_icon:
-
         button.setCheckable(True)
 
-        def changeIcon(button=button,
-                       icon=icon,
-                       toggle_icon=toggle_icon,
-                       size=icon_size):
+        def changeIcon(
+            button=button, icon=icon, toggle_icon=toggle_icon, size=icon_size
+        ):
             if button.isChecked():
                 button.setIcon(pyqt.get_icon(toggle_icon, size))
             else:
@@ -426,3 +425,79 @@ def create_button(size=17,
         button.clicked.connect(changeIcon)
 
     return button
+
+
+#########################
+# Widget Settings Manager
+#########################
+
+
+class WidgetSettingsManager(QtCore.QSettings):
+    widget_map = {
+        "QCheckBox": ("isChecked", "setChecked", bool, False),
+        "QComboBox": ("currentIndex", "setCurrentIndex", int, 0),
+        "QLineEdit": ("text", "setText", str, ""),
+        "QListWidget": (None, None, str, ""),
+    }
+
+    def __init__(self, ui_name, parent=None):
+        super(WidgetSettingsManager, self).__init__(parent)
+        self.settings = QtCore.QSettings(
+            QtCore.QSettings.IniFormat,
+            QtCore.QSettings.UserScope,
+            "mcsGear",
+            ui_name,
+        )
+
+    def _get_listwidget_item_names(self, listwidget):
+        items = [listwidget.item(i).text() for i in range(listwidget.count())]
+        item_string = ",".join(items)
+        return item_string
+
+    def _add_listwidget_items(self, name, listwidget):
+        value = self.settings.value(name, type=str)
+        if not value or value == "0":
+            return
+        items = value.split(",")
+        current_items = self._get_listwidget_item_names(listwidget)
+        for item in items:
+            if item not in current_items:
+                listwidget.addItem(item)
+
+    def save_ui_state(self, widget_dict):
+        for name, widget in widget_dict.items():
+            class_name = widget.__class__.__name__
+            if class_name not in self.widget_map:
+                continue
+            if class_name == "QListWidget":
+                value = self._get_listwidget_item_names(widget)
+                self.settings.setValue(name, value)
+                continue
+            getter, _, _, _ = self.widget_map.get(class_name)
+            if not getter:
+                return
+            get_function = getattr(widget, getter)
+            value = get_function()
+            if value is not None:
+                self.settings.setValue(name, value)
+
+    def load_ui_state(self, widget_dict, reset=False):
+        for name, widget in widget_dict.items():
+            class_name = widget.__class__.__name__
+            if class_name not in self.widget_map:
+                continue
+            if class_name == "QListWidget":
+                self._add_listwidget_items(name, widget)
+                continue
+            _, setter, dtype, default_value = self.widget_map.get(class_name)
+            if not setter:
+                return
+            value = self.settings.value(name, type=dtype)
+            if reset:
+                value = default_value
+            if value is not None:
+                set_function = getattr(widget, setter)
+                try:
+                    set_function(value)
+                except Exception as e:
+                    print(e)
